backend/backend/requirements_new.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1